added logo, removed some text, added exponent
I made a logo using https://textkool.com/en/test-ascii-art-generator?text=badcalc, I eventually decided to remove part of the prompt that said along the lines of please choose a calculator type. I thought having that under the logo looked kinda bad and I didn't care enough about it to put it centered or around the displayed calculator modes. In addition to that I wanted to add the ability to calculate exponents to the calculator, as you can see I add that as well.