comments and more
commented the code, I can now take that of the todo list. Also made added a check to tell you that you need to put a valid calcmode if the user puts anything lower than 1 in addition to the already present check if the user puts anything higher than 6 for the calcmode. Now if you enter anything higher than 6 or lower than one it will output "Please enter a valid calculator action"