added trig functions, still more to add
I added sine cosine and tangent to the calculator I plan to add the ability to get the sine, cosine and tangent of angle theta as well, I have to figure out if i want to find i way to do determine if it is sine of theta, cosine of theta , tangent of theta or just add individual functions for each one. I will likely try each one and use the one that uses the least amount of lines of code.